+++ poptorch/source/PoptorchSymbols.cpp
@@ -33,6 +33,7 @@ c10::Symbol nll_loss_nd;                 // NOLINT
 c10::Symbol cross_entropy_loss;          // NOLINT
 c10::Symbol new_ones;                    // NOLINT
 c10::Symbol new_zeros;                   // NOLINT
+c10::Symbol new_full;                    // NOLINT
 c10::Symbol numpy_T;                     // NOLINT
 c10::Symbol col2im;                      // NOLINT
 c10::Symbol im2col;                      // NOLINT
@@ -73,6 +74,7 @@ static void initializeAtenSymbols() {
   SYMBOL_INIT(aten, cross_entropy_loss);
   SYMBOL_INIT(aten, new_ones);
   SYMBOL_INIT(aten, new_zeros);
+  SYMBOL_INIT(aten, new_full);
   SYMBOL_INIT(aten, numpy_T);
   SYMBOL_INIT(aten, col2im);
   SYMBOL_INIT(aten, im2col);