+++ poptorch/source/PoptorchSymbols.hpp
@@ -32,6 +32,7 @@ extern c10::Symbol nll_loss_nd;                 // NOLINT
 extern c10::Symbol cross_entropy_loss;          // NOLINT
 extern c10::Symbol new_ones;                    // NOLINT
 extern c10::Symbol new_zeros;                   // NOLINT
+extern c10::Symbol new_full;                    // NOLINT
 extern c10::Symbol numpy_T;                     // NOLINT
 extern c10::Symbol col2im;                      // NOLINT
 extern c10::Symbol im2col;                      // NOLINT